+++ docs/changelog.md
@@ -6,6 +6,9 @@ Requires libvips v8.15.2
 
 ### v0.33.5 - TBD
 
+* Add `pageHeight` and `pages` to response of multi-page output.
+  [#3411](https://github.com/lovell/sharp/issues/3411)
+
 * Ensure option to force use of a globally-installed libvips works correctly.
   [#4111](https://github.com/lovell/sharp/pull/4111)
   [@project0](https://github.com/project0)